fix: improved hash-key for openml cache
Previously, the whole solutions.qmd contributed to the hash-key
for the openml cache. This led to issues when many people
were working on the solutions, because the branches did not
key a cache-hit and only a restore-hit. This caused all the
branches in the solutions to upload the cache. This took
away space for actually relevant caches