fix(ci): include mermaid diagrams as png and svg (technical)
For a while, we have suffered from random timeouts when rendering the book.
The log-output from GHA during the "Render book" stage just ended with:

|.............................................| 100% output file: advanced_technical_aspects_of_mlr3.knit.md

Error: The operation was canceled.

after hitting the maximum runtime of 6 hours.
(the success rate was around 50/50).

When rendering with the --execute-debug flag
more log-output was given.
The log output at the end of rendering the technical chapter (pdf) was:

|
|.............................................| 100%

output file: advanced_technical_aspects_of_mlr3.knit.md

[knitr engine]: writing results
[knitr engine]: exiting
Error: The operation was canceled.

for the other chapters when rendering to pdf, the output was

|............................................| 100%

output file: preprocessing.knit.md

[knitr engine]: writing results
[knitr engine]: exiting
[knitr engine]: postprocess
[knitr engine]: writing results
[knitr engine]: exiting

--> something with the postprocessing went wrong and the bug was identified to be in the technical chapter

The problem was NOT the large-scale benchmarking chapter

https://github.com/quarto-dev/quarto-actions/issues/45#issuecomment-1798392018